# frontend/openapi.json
{
  "components": {
    "schemas": {
      "AnalysisDetail": {
        "additionalProperties": false,
        "properties": {
          "provenance": {
            "$ref": "#/components/schemas/Provenance"
          },
          "report": {
            "$ref": "#/components/schemas/DRReadinessReport"
          },
          "stale": {
            "description": "Older than HISTORY_STALE_AFTER_HOURS; health may differ now",
            "title": "Stale",
            "type": "boolean"
          },
          "summary": {
            "$ref": "#/components/schemas/AnalysisSummary"
          }
        },
        "required": [
          "summary",
          "report",
          "provenance",
          "stale"
        ],
        "title": "AnalysisDetail",
        "type": "object"
      },
      "AnalysisPage": {
        "additionalProperties": false,
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/AnalysisSummary"
            },
            "title": "Items",
            "type": "array"
          },
          "nextBefore": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Pass as `before` for the next page; null on the last page",
            "title": "Nextbefore"
          }
        },
        "required": [
          "items"
        ],
        "title": "AnalysisPage",
        "type": "object"
      },
...
      "AnalysisSummary": {
        "additionalProperties": false,
        "properties": {
          "aiAnalysisAvailable": {
            "title": "Aianalysisavailable",
            "type": "boolean"
          },
          "completedAt": {
            "format": "date-time",
            "title": "Completedat",
            "type": "string"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "executionCount": {
            "minimum": 0.0,
            "title": "Executioncount",
            "type": "integer"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "inventoryLabel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Inventorylabel"
          },
          "owner": {
            "title": "Owner",
            "type": "string"
          },
          "riskLevel": {
            "$ref": "#/components/schemas/RiskLevel"
          },
          "riskScore": {
            "title": "Riskscore",
            "type": "integer"
          },
          "rtoFeasible": {
            "title": "Rtofeasible",
            "type": "boolean"
          },
          "runbookLabel": {
            "title": "Runbooklabel",
            "type": "string"
          },
          "serviceName": {
            "title": "Servicename",
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/AnalysisSource"
          }
        },
        "required": [
          "id",
          "serviceName",
          "owner",
          "source",
          "createdAt",
          "completedAt",
          "runbookLabel",
          "inventoryLabel",
          "riskScore",
          "riskLevel",
          "rtoFeasible",
          "aiAnalysisAvailable",
          "executionCount"
        ],
        "title": "AnalysisSummary",
        "type": "object"
      },
...
      "DRReadinessReport": {
        "additionalProperties": false,
        "properties": {
          "aiAnalysisAvailable": {
            "default": true,
            "title": "Aianalysisavailable",
            "type": "boolean"
          },
          "aiNote": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ainote"
          },
          "dependencyHealth": {
            "items": {
              "$ref": "#/components/schemas/DependencyHealth"
            },
            "title": "Dependencyhealth",
            "type": "array"
          },
          "executionPlan": {
            "items": {
              "$ref": "#/components/schemas/ExecutionPhase"
            },
            "title": "Executionplan",
            "type": "array"
          },
          "gapAnalysis": {
            "items": {
              "$ref": "#/components/schemas/Gap"
            },
            "title": "Gapanalysis",
            "type": "array"
          },
          "historicalInsights": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ServiceHistory"
              },
              {
                "type": "null"
              }
            ]
          },
          "meta": {
            "$ref": "#/components/schemas/ReportMeta"
          },
          "riskLevel": {
            "$ref": "#/components/schemas/RiskLevel",
            "readOnly": true
          },
          "riskScore": {
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Riskscore",
            "type": "integer"
          },
          "rtoAnalysis": {
            "$ref": "#/components/schemas/RtoAnalysis"
          },
          "serviceSummary": {
            "$ref": "#/components/schemas/ServiceSummary"
          },
          "singlePointsOfFailure": {
            "items": {
              "$ref": "#/components/schemas/SinglePointOfFailure"
            },
            "title": "Singlepointsoffailure",
            "type": "array"
          },
          "suggestions": {
            "items": {
              "$ref": "#/components/schemas/Suggestion"
            },
            "title": "Suggestions",
            "type": "array"
          },
          "summary": {
            "title": "Summary",
            "type": "string"
          }
        },
        "required": [
          "meta",
          "serviceSummary",
          "riskScore",
          "rtoAnalysis",
          "summary",
          "riskLevel"
        ],
        "title": "DRReadinessReport",
        "type": "object"
      },
...
      "ErrorBody": {
        "additionalProperties": false,
        "description": "The shared error shape: `{error, code, details?}`.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "details": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          },
          "error": {
            "title": "Error",
            "type": "string"
          }
        },
        "required": [
          "error",
          "code"
        ],
        "title": "ErrorBody",
        "type": "object"
      },
...
      "ExecutionSummary": {
        "additionalProperties": false,
        "properties": {
          "analysisJobId": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analysisjobid"
          },
          "createdAt": {
            "format": "date-time",
            "title": "Createdat",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "mode": {
            "$ref": "#/components/schemas/ExecutionMode"
          },
          "runbookLabel": {
            "title": "Runbooklabel",
            "type": "string"
          },
          "startedBy": {
            "title": "Startedby",
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ExecutionState"
          },
          "steps": {
            "title": "Steps",
            "type": "integer"
          },
          "updatedAt": {
            "format": "date-time",
            "title": "Updatedat",
            "type": "string"
          }
        },
        "required": [
          "id",
          "analysisJobId",
          "state",
          "mode",
          "runbookLabel",
          "startedBy",
          "createdAt",
          "updatedAt",
          "steps"
        ],
        "title": "ExecutionSummary",
        "type": "object"
      },
...
      "Provenance": {
        "additionalProperties": false,
        "description": "How a report was produced; never the API key or base URL.",
        "properties": {
          "agentVersion": {
            "title": "Agentversion",
            "type": "string"
          },
          "llmModel": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Llmmodel"
          },
          "llmProvider": {
            "title": "Llmprovider",
            "type": "string"
          },
          "promptVersion": {
            "title": "Promptversion",
            "type": "string"
          }
        },
        "required": [
          "llmProvider",
          "promptVersion",
          "agentVersion"
        ],
        "title": "Provenance",
        "type": "object"
      },
...
      "RiskLevel": {
        "enum": [
          "LOW",
          "MEDIUM",
          "HIGH",
          "CRITICAL"
        ],
        "title": "RiskLevel",
        "type": "string"
      },
...
      "ServiceHistory": {
        "additionalProperties": false,
        "description": "Facts computed by code from stored analyses and live executions only.",
        "properties": {
          "analysesConsidered": {
            "minimum": 0.0,
            "title": "Analysesconsidered",
            "type": "integer"
          },
          "dependencies": {
            "items": {
              "$ref": "#/components/schemas/DependencyTrend"
            },
            "title": "Dependencies",
            "type": "array"
          },
          "dryRuns": {
            "description": "Counted but never measured",
            "minimum": 0.0,
            "title": "Dryruns",
            "type": "integer"
          },
          "executions": {
            "description": "Newest first",
            "items": {
              "$ref": "#/components/schemas/ExecutionOutcome"
            },
            "title": "Executions",
            "type": "array"
          },
          "liveRuns": {
            "minimum": 0.0,
            "title": "Liveruns",
            "type": "integer"
          },
          "riskScores": {
            "description": "Oldest first",
            "items": {
              "type": "integer"
            },
            "title": "Riskscores",
            "type": "array"
          },
          "serviceName": {
            "title": "Servicename",
            "type": "string"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/StepHistory"
            },
            "title": "Steps",
            "type": "array"
          }
        },
        "required": [
          "serviceName",
          "analysesConsidered",
          "liveRuns",
          "dryRuns"
        ],
        "title": "ServiceHistory",
        "type": "object"
      },
...
      "StoredExecution": {
        "additionalProperties": false,
        "description": "A past execution of an analysis, read-only: its steps and its verified audit log.",
        "properties": {
          "audit": {
            "$ref": "#/components/schemas/AuditView"
          },
          "execution": {
            "$ref": "#/components/schemas/Execution"
          }
        },
        "required": [
          "execution",
          "audit"
        ],
        "title": "StoredExecution",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Parse DR runbooks, check dependency health and produce readiness reports.",
    "title": "DR Readiness Agent API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
    "/api/v1/analyses": {
      "get": {
        "description": "Stored analyses, newest first.",
        "operationId": "list_analyses_api_v1_analyses_get",
        "parameters": [
          {
            "in": "query",
            "name": "service",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "minLength": 1,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service"
            }
          },
          {
            "in": "query",
            "name": "riskLevel",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/RiskLevel"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Risklevel"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 100,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "description": "`nextBefore` of the last page",
            "in": "query",
            "name": "before",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "`nextBefore` of the last page",
              "title": "Before"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisPage"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "List Analyses",
        "tags": [
          "history"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}": {
      "delete": {
        "description": "`409` when executions refer to it: their audit trail must keep its source.",
        "operationId": "delete_analysis_api_v1_analyses__analysis_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Delete Analysis",
        "tags": [
          "history"
        ]
      },
      "get": {
        "operationId": "get_analysis_api_v1_analyses__analysis_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnalysisDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Analysis",
        "tags": [
          "history"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}/executions": {
      "get": {
        "description": "Executions created from this analysis (also readable while execution is off).",
        "operationId": "get_analysis_executions_api_v1_analyses__analysis_id__executions_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ExecutionSummary"
                  },
                  "title": "Response Get Analysis Executions Api V1 Analyses  Analysis Id  Executions Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Analysis Executions",
        "tags": [
          "history"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}/executions/{execution_id}": {
      "get": {
        "description": "One execution of this analysis with its steps and its verified audit log (read-only).",
        "operationId": "get_analysis_execution_api_v1_analyses__analysis_id__executions__execution_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "execution_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Execution Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StoredExecution"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Analysis Execution",
        "tags": [
          "history"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}/report.html": {
      "get": {
        "operationId": "get_analysis_report_html_api_v1_analyses__analysis_id__report_html_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/html": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "text/html": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Analysis Report Html",
        "tags": [
          "history"
        ]
      }
    },
    "/api/v1/analyses/{analysis_id}/runbook": {
      "get": {
        "description": "The runbook Markdown exactly as analyzed, as a text download.",
        "operationId": "get_analysis_runbook_api_v1_analyses__analysis_id__runbook_get",
        "parameters": [
          {
            "in": "path",
            "name": "analysis_id",
            "required": true,
            "schema": {
              "pattern": "^[A-Za-z0-9_-]{1,64}$",
              "title": "Analysis Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Analysis Runbook",
        "tags": [
          "history"
        ]
      }
    },
...
    "/api/v1/services/{service_name}/history": {
      "get": {
        "description": "Measured facts from this service's stored analyses and finished live runs (ADR 0009).\n\nSteps are keyed by fingerprint; a report's `historicalInsights` has them matched\nto that runbook's step numbers instead.",
        "operationId": "get_service_history_api_v1_services__service_name__history_get",
        "parameters": [
          {
            "in": "path",
            "name": "service_name",
            "required": true,
            "schema": {
              "maxLength": 200,
              "minLength": 1,
              "title": "Service Name",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ServiceHistory"
                }
              }
            },
            "description": "Successful Response"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Not Found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorBody"
                }
              }
            },
            "description": "Unprocessable Entity"
          }
        },
        "summary": "Get Service History",
        "tags": [
          "history"
        ]
      }
    },
...
  }
}